# storage/swagger/http.json
{"openapi":"3.0.0","paths":{"\/logout":{"post":{"tags":["v1\/"],"description":"退出登录","operationId":"729b976e29ad68934e0f3412908fbe0c","parameters":[{"name":"authorization","in":"header","description":"token","required":true,"schema":{"type":"string"},"example":"Bearer 71e4a1aa3dd12e43164b0e407457f6b4"}],"responses":{"200":{"description":"响应参数","content":{"application\/json":{"schema":{"properties":{"code":{"description":"状态码","type":"integer","rules":null},"message":{"description":"消息","type":"string","rules":null},"data":{"description":"数据","type":"object","rules":null}},"type":"object"}}}}}}},"\/login":{"post":{"tags":["v1\/"],"description":"登录","operationId":"6e7b3666536e482be1834c2b58dbbf38","requestBody":{"description":"请求参数","content":{"application\/x-www-form-urlencoded":{"schema":{"required":["email","password"],"properties":{"email":{"description":"邮箱","type":"string","rules":null},"password":{"description":"密码","type":"string","rules":null}},"type":"object"}}},"_content":[{"mediaType":"application\/x-www-form-urlencoded","schema":{"required":["email","password"],"properties":{"email":{"description":"邮箱","type":"string","rules":null},"password":{"description":"密码","type":"string","rules":null}},"type":"object"}}]},"responses":{"200":{"description":"响应参数","content":{"application\/json":{"schema":{"properties":{"code":{"description":"状态码","type":"integer","rules":null},"message":{"description":"消息","type":"string","rules":null},"data":{"description":"数据","properties":{"name":{"description":"用户名","type":"string","rules":null},"email":{"description":"邮箱","type":"string","rules":null},"token":{"description":"token","type":"string","rules":null}},"type":"object","rules":null}},"type":"object"}}}}}}},"\/send-captcha":{"post":{"tags":["v1\/"],"description":"发送验证码","operationId":"9f6d50a70820e65388f09b037718b272","requestBody":{"description":"请求参数","content":{"application\/x-www-form-urlencoded":{"schema":{"required":["email"],"properties":{"email":{"description":"邮箱","type":"string","rules":null}},"type":"object"}}},"_content":[{"mediaType":"application\/x-www-form-urlencoded","schema":{"required":["email"],"properties":{"email":{"description":"邮箱","type":"string","rules":null}},"type":"object"}}]},"responses":{"200":{"description":"响应参数","content":{"application\/json":{"schema":{"properties":{"code":{"description":"状态码","type":"integer","rules":null},"message":{"description":"消息","type":"string","rules":null},"data":{"description":"数据","type":"object","rules":null}},"type":"object"}}}}}}},"\/register":{"post":{"tags":["v1\/"],"description":"邮箱注册","operationId":"15b569735581a6652e2d19033cfcf9e2","requestBody":{"description":"请求参数","content":{"application\/x-www-form-urlencoded":{"schema":{"required":["email","password","captcha"],"properties":{"email":{"description":"邮箱","type":"string","rules":null},"password":{"description":"密码","type":"string","rules":null},"captcha":{"description":"验证码","type":"string","rules":null}},"type":"object"}}},"_content":[{"mediaType":"application\/x-www-form-urlencoded","schema":{"required":["email","password","captcha"],"properties":{"email":{"description":"邮箱","type":"string","rules":null},"password":{"description":"密码","type":"string","rules":null},"captcha":{"description":"验证码","type":"string","rules":null}},"type":"object"}}]},"responses":{"200":{"description":"响应参数","content":{"application\/json":{"schema":{"properties":{"code":{"description":"状态码","type":"integer","rules":null},"message":{"description":"消息","type":"string","rules":null},"data":{"description":"数据","properties":{"name":{"description":"用户名","type":"string","rules":null},"email":{"description":"邮箱","type":"string","rules":null},"token":{"description":"token","type":"string","rules":null}},"type":"object","rules":null}},"type":"object"}}}}}}}}}